feat(docs): add wishlist functionality and update related documentation
