+++ apps/docs/src/frontends-recipes/account/index.md
@@ -9,3 +9,5 @@ nav:
 Recipes for customer session and account flows.
 
 <PageRef page="login.html" title="Login" sub="Understand the customer login flow, Store API operations, generated types, context refresh, cart refresh, and error handling." />
+
+<PageRef page="wishlist.html" title="Wishlist" sub="Understand the customer wishlist as a paginated product search, the Store API wishlist operations, the shared composable state, and what one page of results does not tell you." />
